Change timeslot time property to long Allow all origins Mark received timeslots from webclimber always as bookable

+++ BoulderingTimes.Api/Models/TimeSlot.cs
@@ -3,9 +3,9 @@
     public class TimeSlot
     {
         public object[] selector { get; set; }
-        public int bookableFrom { get; set; }
+        public long bookableFrom { get; set; }
         public string state { get; set; }
-        public int bookableUntilDuration { get; set; }
+        public long bookableUntilDuration { get; set; }
         public int? minCourseParticipantCount { get; set; }
         public int? maxCourseParticipantCount { get; set; }
         public int? currentCourseParticipantCount { get; set; }
